Remove old dashboard.
