+++ dashboard-feature/src/main/java/eu/europa/ec/dashboardfeature/router/Graph.kt
@@ -22,7 +22,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
 import androidx.navigation.navDeepLink
 import eu.europa.ec.dashboardfeature.BuildConfig
-import eu.europa.ec.dashboardfeature.ui.dashboard_new.DashboardScreenNew
+import eu.europa.ec.dashboardfeature.ui.dashboard.DashboardScreen
 import eu.europa.ec.dashboardfeature.ui.sign.DocumentSignScreen
 import eu.europa.ec.uilogic.navigation.DashboardScreens
 import eu.europa.ec.uilogic.navigation.ModuleRoute
@@ -42,16 +42,13 @@ fun NavGraphBuilder.featureDashboardGraph(navController: NavController) {
                 }
             )
         ) {
-            //TODO change this accordingly.
-            DashboardScreenNew(
+            DashboardScreen(
                 hostNavController = navController,
                 viewModel = koinViewModel(),
                 documentsViewModel = koinViewModel(),
                 homeViewModel = koinViewModel(),
                 transactionsViewModel = koinViewModel()
             )
-
-            //DashboardScreen(navController, koinViewModel())
         }
 
         composable(
