+++ dashboard-feature/src/main/java/eu/europa/ec/dashboardfeature/ui/documents/DocumentsScreen.kt
@@ -89,8 +89,8 @@ import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 
-typealias DashboardEvent = eu.europa.ec.dashboardfeature.ui.dashboard_new.Event
-typealias ShowSideMenuEvent = eu.europa.ec.dashboardfeature.ui.dashboard_new.Event.SideMenu.Show
+typealias DashboardEvent = eu.europa.ec.dashboardfeature.ui.dashboard.Event
+typealias ShowSideMenuEvent = eu.europa.ec.dashboardfeature.ui.dashboard.Event.SideMenu.Show
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
